Refactor lib (#292)
Refactor imports
