+++ src/components/Layout/Header/Header.tsx
@@ -1,10 +1,9 @@
 import { useMatch } from 'react-router-dom';
-import { Button } from 'components/Button';
-import { MxLink } from 'components/MxLink';
+import { Button, MxLink } from 'components';
 import { environment } from 'config';
-import { useGetIsLoggedIn, logout } from 'utils/sdkDapp';
-import MultiversXLogo from '../../../assets/img/multiversx-logo.svg?react';
+import { useGetIsLoggedIn, logout } from 'lib';
 import { RouteNamesEnum } from 'localConstants';
+import MultiversXLogo from '../../../assets/img/multiversx-logo.svg?react';
 
 const callbackUrl = `${window.location.origin}/unlock`;
 const onRedirect = undefined; // use this to redirect with useNavigate to a specific page after logout
