+++ src/helpers/signAndSendTransactions.ts
@@ -1,7 +1,6 @@
+import { refreshAccount, sendTransactions } from 'lib';
 import { isSafari } from 'localConstants';
 import { Transaction, TransactionsDisplayInfoType } from 'types';
-
-import { refreshAccount, sendTransactions } from 'utils/sdkDapp';
 
 type SignAndSendTransactionsProps = {
   transactions: Transaction[];
