+++ src/lib/sdkCore/sdkCore.types.ts
