+++ src/lib/sdkDapp/sdkDapp.constants.ts
@@ -1,0 +1,8 @@
+export { TRANSACTIONS_ENDPOINT } from '@multiversx/sdk-dapp/apiCalls/endpoints';
+export {
+  DECIMALS,
+  GAS_PRICE,
+  GAS_LIMIT,
+  EXTRA_GAS_LIMIT_GUARDED_TX,
+  VERSION
+} from '@multiversx/sdk-dapp/constants';
