+++ src/lib/sdkDapp/sdkDapp.helpers.ts
@@ -1,0 +1,22 @@
+export { addressIsValid } from '@multiversx/sdk-dapp/utils/account/addressIsValid';
+export { getInterpretedTransaction } from '@multiversx/sdk-dapp/utils/transactions/getInterpretedTransaction';
+export { formatAmount } from '@multiversx/sdk-dapp/utils/operations/formatAmount';
+export { getIsProviderEqualTo } from '@multiversx/sdk-dapp/utils/account/getIsProviderEqualTo';
+export { refreshAccount } from '@multiversx/sdk-dapp/utils/account/refreshAccount';
+export {
+  deleteTransactionToast,
+  removeAllSignedTransactions,
+  removeAllTransactionsToSign
+} from '@multiversx/sdk-dapp/services/transactions/clearTransactions';
+export {
+  setTransactionsDisplayInfoState,
+  setTransactionsToSignedState
+} from '@multiversx/sdk-dapp/services/transactions/updateSignedTransactions';
+export { verifyMessage } from '@multiversx/sdk-dapp/hooks/signMessage/verifyMessage';
+export { sendBatchTransactions } from '@multiversx/sdk-dapp/services/transactions/sendBatchTransactions';
+export { getTransactions } from '@multiversx/sdk-dapp/apiCalls/transactions/getTransactions';
+export { sendTransactions } from '@multiversx/sdk-dapp/services/transactions/sendTransactions';
+export { logout } from '@multiversx/sdk-dapp/utils/logout';
+export { signTransactions } from '@multiversx/sdk-dapp/services/transactions/signTransactions';
+export { trimUsernameDomain } from '@multiversx/sdk-dapp/hooks/account/helpers';
+export { newTransaction } from '@multiversx/sdk-dapp/models';
