+++ src/lib/sdkDapp/sdkDapp.hooks.ts
@@ -1,18 +1,3 @@
-export { addressIsValid } from '@multiversx/sdk-dapp/utils/account/addressIsValid';
-export { getInterpretedTransaction } from '@multiversx/sdk-dapp/utils/transactions/getInterpretedTransaction';
-export { formatAmount } from '@multiversx/sdk-dapp/utils/operations/formatAmount';
-export { getIsProviderEqualTo } from '@multiversx/sdk-dapp/utils/account/getIsProviderEqualTo';
-export { refreshAccount } from '@multiversx/sdk-dapp/utils/account/refreshAccount';
-export {
-  deleteTransactionToast,
-  removeAllSignedTransactions,
-  removeAllTransactionsToSign
-} from '@multiversx/sdk-dapp/services/transactions/clearTransactions';
-export {
-  setTransactionsDisplayInfoState,
-  setTransactionsToSignedState
-} from '@multiversx/sdk-dapp/services/transactions/updateSignedTransactions';
-export { sendBatchTransactions } from '@multiversx/sdk-dapp/services/transactions/sendBatchTransactions';
 export { useGetAccount } from '@multiversx/sdk-dapp/hooks/account/useGetAccount';
 export { useGetAccountInfo } from '@multiversx/sdk-dapp/hooks/account/useGetAccountInfo';
 export { useGetLoginInfo } from '@multiversx/sdk-dapp/hooks/account/useGetLoginInfo';
@@ -25,7 +10,6 @@ export { useGetActiveTransactionsStatus } from '@multiversx/sdk-dapp/hooks/trans
 export { useGetPendingTransactions } from '@multiversx/sdk-dapp/hooks/transactions/useGetPendingTransactions';
 export { useGetNetworkConfig } from '@multiversx/sdk-dapp/hooks/useGetNetworkConfig';
 export { useTrackTransactionStatus } from '@multiversx/sdk-dapp/hooks/transactions/useTrackTransactionStatus';
-export { verifyMessage } from '@multiversx/sdk-dapp/hooks/signMessage/verifyMessage';
 export { useSendBatchTransactions } from '@multiversx/sdk-dapp/hooks/transactions/batch/useSendBatchTransactions';
 export { useGetBatches } from '@multiversx/sdk-dapp/hooks/transactions/batch/useGetBatches';
 export { useCheckBatch } from '@multiversx/sdk-dapp/hooks/transactions/batch/tracker/useCheckBatch';
@@ -34,17 +18,3 @@ export { useBatchTransactionsTracker } from '@multiversx/sdk-dapp/hooks/transact
 export { useGetSignedTransactions } from '@multiversx/sdk-dapp/hooks/transactions/useGetSignedTransactions';
 export { useGetAccountProvider } from '@multiversx/sdk-dapp/hooks/account/useGetAccountProvider';
 export { useTransactionsTracker } from '@multiversx/sdk-dapp/hooks/transactions/useTransactionsTracker';
-export { getTransactions } from '@multiversx/sdk-dapp/apiCalls/transactions/getTransactions';
-export { sendTransactions } from '@multiversx/sdk-dapp/services/transactions/sendTransactions';
-export { logout } from '@multiversx/sdk-dapp/utils/logout';
-export { signTransactions } from '@multiversx/sdk-dapp/services/transactions/signTransactions';
-export { trimUsernameDomain } from '@multiversx/sdk-dapp/hooks/account/helpers';
-export { newTransaction } from '@multiversx/sdk-dapp/models';
-export { TRANSACTIONS_ENDPOINT } from '@multiversx/sdk-dapp/apiCalls/endpoints';
-export {
-  DECIMALS,
-  GAS_PRICE,
-  GAS_LIMIT,
-  EXTRA_GAS_LIMIT_GUARDED_TX,
-  VERSION
-} from '@multiversx/sdk-dapp/constants';
