+++ src/pages/Dashboard/widgets/BatchTransactions/helpers/signAndAutoSendBatchTransactions.ts
@@ -1,5 +1,8 @@
 import { SessionEnum, isSafari } from 'localConstants/session';
-import { refreshAccount, sendBatchTransactions } from 'utils/sdkDapp';
+import {
+  refreshAccount,
+  sendBatchTransactions
+} from 'lib';
 import { getBatchTransactions } from '../helpers';
 import { SendTransactionProps } from '../types';
