+++ src/pages/Dashboard/widgets/PingPongService/PingPongService.tsx
@@ -2,16 +2,19 @@ import { useEffect, useState } from 'react';
 import { faArrowUp, faArrowDown } from '@fortawesome/free-solid-svg-icons';
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 import moment from 'moment';
-import { Button } from 'components/Button';
-import { ContractAddress } from 'components/ContractAddress';
-import { Label } from 'components/Label';
-import { MissingNativeAuthError } from 'components/MissingNativeAuthError';
-import { OutputContainer, PingPongOutput } from 'components/OutputContainer';
+import {
+  Button,
+  ContractAddress,
+  Label,
+  MissingNativeAuthError,
+  OutputContainer,
+  PingPongOutput
+} from 'components';
 import { getCountdownSeconds, setTimeRemaining } from 'helpers';
 import { useSendPingPongTransaction } from 'hooks';
+import { useGetLoginInfo, useGetPendingTransactions } from 'lib';
 import { SessionEnum } from 'localConstants';
 import { SignedTransactionType, WidgetProps } from 'types';
-import { useGetLoginInfo, useGetPendingTransactions } from 'utils/sdkDapp';
 import {
   useGetTimeToPong,
   useGetPingTransaction,
