+++ src/pages/Dashboard/widgets/PingPongService/hooks/useGetPongTransaction.ts
@@ -1,6 +1,6 @@
 import axios from 'axios';
 import { API_URL } from 'config';
-import { Transaction } from 'types/sdkCore.types';
+import { Transaction } from 'lib';
 
 export const useGetPongTransaction = () => {
   return async () => {
