+++ src/pages/Unlock/Unlock.tsx
@@ -8,14 +8,14 @@ import {
 } from '@multiversx/sdk-dapp/UI';
 import { IframeLoginTypes } from '@multiversx/sdk-web-wallet-iframe-provider/out/constants';
 import { useNavigate } from 'react-router-dom';
+import { nativeAuth } from 'config';
+import { useWindowSize } from 'hooks';
 import {
   ExtensionLoginButton,
   LedgerLoginButton,
   OperaWalletLoginButton,
   WalletConnectLoginButton
-} from 'components/sdkDapp';
-import { nativeAuth } from 'config';
-import { useWindowSize } from 'hooks';
+} from 'lib';
 import { RouteNamesEnum } from 'localConstants';
 import {
   IframeButton,
