+++ src/pages/Unlock/components/WebWalletLoginWrapper.tsx
@@ -1,7 +1,6 @@
-import React from 'react';
 import { useWebWalletLogin } from '@multiversx/sdk-dapp/hooks/login/useWebWalletLogin';
 import { WebWalletLoginButtonPropsType } from '@multiversx/sdk-dapp/UI/webWallet/WebWalletLoginButton/WebWalletLoginButton';
-import { CrossWindowLoginButton } from 'components/sdkDapp';
+import { CrossWindowLoginButton } from 'lib';
 import { DualLoginButton } from './DualLoginButton';
 
 export const WebWalletLoginWrapper = ({
