+++ src/types/index.ts
@@ -2,5 +2,5 @@ export * from './profile.types';
 export * from './widget.types';
 export * from './pingPong.types';
 export * from './transaction.types';
-export * from './sdkDapp.types';
-export * from './sdkCore.types';
+export * from '../lib/sdkDapp/sdkDapp.types';
+export * from '../lib/sdkCore/sdkCore.types';
