+++ src/utils/index.ts
@@ -1,1 +1,3 @@
 export * from './getCallbackRoute';
+export * from './getChainId';
+export * from './smartContract';
