+++ src/wrappers/AuthRedirectWrapper/AuthRedirectWrapper.tsx
@@ -1,12 +1,7 @@
 import { PropsWithChildren, useEffect } from 'react';
-import {
-  matchRoutes,
-  useLocation,
-  useNavigate,
-  matchPath
-} from 'react-router-dom';
+import { useLocation, useNavigate, matchPath } from 'react-router-dom';
+import { useGetIsLoggedIn } from 'lib';
 import { RouteNamesEnum } from 'localConstants';
-import { useGetIsLoggedIn } from 'utils/sdkDapp';
 import { routes } from 'routes';
 
 export const AuthRedirectWrapper = ({ children }: PropsWithChildren) => {
